+++ src/test/java/step2/domain/BowLingScoreViewFormatTest.java
@@ -1,0 +1,35 @@
+package step2.domain;
+
+import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import step2.view.BowLingScoreView;
+
+import java.util.Arrays;
+import java.util.List;
+
+class BowLingScoreViewTest {
+    @DisplayName("spair 결과 테스트")
+    @Test
+    void strikeViewTest() {
+        List<Integer> bowlingScores = Arrays.asList(10);
+        List<String> actual = BowLingScoreView.transferBowlingScoreViewFormat(bowlingScores);
+        Assertions.assertThat(actual.get(0)).isEqualTo("X");
+    }
+
+    @DisplayName("spair 결과 테스트")
+    @Test
+    void spairViewTest() {
+        List<Integer> bowlingScores = Arrays.asList(4,6);
+        List<String> actual = BowLingScoreView.transferBowlingScoreViewFormat(bowlingScores);
+        Assertions.assertThat(actual.get(1)).isEqualTo("/");
+    }
+
+    @DisplayName("score 결과 테스트")
+    @Test
+    void scoreViewTest() {
+        List<Integer> bowlingScores = Arrays.asList(4,6);
+        List<String> actual = BowLingScoreView.transferBowlingScoreViewFormat(bowlingScores);
+        Assertions.assertThat(actual.get(0)).isEqualTo("4");
+    }
+}
