--- src/test/java/step2/domain/BowLingScoreViewTest.java
@@ -1,4 +1,0 @@
-import static org.junit.jupiter.api.Assertions.*;
-class BowLingScoreViewTest {
-  
-}
